ml/requirements.txt
# Data processing
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1

# Machine Learning
scikit-learn==1.2.2
xgboost==1.7.5
lightgbm==3.3.5
shap==0.41.0
lime==0.2.0.1
optuna==3.2.0

# Deep Learning (optional)
tensorflow==2.12.0
torch==2.0.1
transformers==4.30.2

# NLP
nltk==3.8.1
spacy==3.5.3
gensim==4.3.1

# API
fastapi==0.98.0
uvicorn==0.22.0
pydantic==1.10.9
python-multipart==0.0.6

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0

# Utilities
joblib==1.2.0
python-dotenv==1.0.0
tqdm==4.65.0

# Testing
pytest==7.3.1